+++ src/app/components/sudoku/board/board.js
