+++ src/app/components/sudoku/container.js
@@ -1,5 +1,5 @@
 'use client'
-import Board from "./board";
+import Board from "./board/board";
 import { useState, useEffect } from "react";
 
 export default function container() {
@@ -8,34 +8,10 @@ export default function container() {
     Array.from({ length: 9 }, () => Array(9).fill('1'))
   );
 
-  function test(){
-    console.log("board", board);
-  }
-
   return ( 
-    <main style={styles.container}>
-      <button onClick={() => test()}> test</button>
+    <main className="sudoku-container">
       <h1 className={'title'}>Sudoku</h1>
       <Board board={board}/>
     </main>
-    // <div style={{
-    //     display: "flex",
-    //     justifyContent: "center",
-    //     alignItems: "center",
-    //     height: "100vh",
-    //     fontSize: "2rem",
-    //     fontWeight: "bold"
-    //   }}>
-    //     Under Development
-    // </div>
-    
  );
-}
-
-const styles = {
-  container: {
-    position: "relative",
-    textAlign: "center",
-    marginTop: "1.5rem",
-  },
 }
