+++ src/app/globals.css
@@ -87,3 +87,8 @@ h1 {
   cursor: default;
 }
 
+.sudoku-container{
+  position: relative;
+  text-align: center;
+  margin-top: 1.5rem;
+}
